towards resultgatherer interface
