+++ CZICheck/checkers/checkerBenabled.h
@@ -17,7 +17,7 @@ public:
 
     CCheckBenabled(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
 };
