+++ CZICheck/checkers/checkerConsecutivePlaneIndices.h
@@ -18,7 +18,7 @@ public:
 
     CCheckConsecutivePlaneIndices(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
 private:
