+++ CZICheck/checkers/checkerConsistentCoordinates.h
@@ -18,7 +18,7 @@ public:
 
     CCheckConsistentCoordinates(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
