+++ CZICheck/checkers/checkerMissingMindex.cpp
@@ -16,7 +16,7 @@ using namespace std;
 
 CCheckMissingMindex::CCheckMissingMindex(
     const std::shared_ptr<libCZI::ICZIReader>& reader,
-    CResultGatherer& result_gatherer,
+    IResultGatherer& result_gatherer,
     const CheckerCreateInfo& additional_info) :
     CCheckerBase(reader, result_gatherer, additional_info)
 {
@@ -47,8 +47,8 @@ void CCheckMissingMindex::RunCheck()
 
     if (count > 0)
     {
-        CResultGatherer::Finding finding(CCheckMissingMindex::kCheckType);
-        finding.severity = CResultGatherer::Severity::Warning;
+        IResultGatherer::Finding finding(CCheckMissingMindex::kCheckType);
+        finding.severity = IResultGatherer::Severity::Warning;
         stringstream ss;
         ss << "There are " << count << " subblocks with no M index.";
         finding.information = ss.str();
