+++ CZICheck/checkers/checkerOverlappingScenes.h
@@ -19,7 +19,7 @@ public:
 
     CCheckOverlappingScenesOnLayer0(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
 private:
