+++ CZICheck/checkers/checkerSamePixeltypePerChannel.h
@@ -20,7 +20,7 @@ public:
 
     CCheckSamePixeltypePerChannel(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
 private:
