+++ CZICheck/checkers/checkerSubBlkBitmapValid.h
@@ -18,7 +18,7 @@ public:
 
     CCheckSubBlkBitmapValid(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
 };
