+++ CZICheck/checkers/checkerSubBlkDirPositions.h
@@ -21,7 +21,7 @@ public:
 
     CCheckSubBlkDirPositions(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
 };
