+++ CZICheck/checkers/checkerTopographyApplianceValidation.h
@@ -49,7 +49,7 @@ public:
     
     CCheckTopographyApplianceMetadata(
         const std::shared_ptr<libCZI::ICZIReader>& reader,
-        CResultGatherer& result_gatherer,
+        IResultGatherer& result_gatherer,
         const CheckerCreateInfo& additional_info);
     void RunCheck() override;
