+++ CZICheck/runchecks.h
@@ -5,6 +5,7 @@
 #pragma once
 #include "cmdlineoptions.h"
 #include "resultgatherer.h"
+#include "resultgathererjson.h"
 #include <memory>
 
 /// This class is responsible for running the checks.
@@ -16,5 +17,5 @@ private:
 public:
     CRunChecks(const CCmdLineOptions& opts, std::shared_ptr<ILog> consoleIo);
 
-    bool Run(CResultGatherer::AggregatedResult& result);
+    bool Run(IResultGatherer::AggregatedResult& result);
 };
